% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire Dataset}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{how much did the fire spread.}
  \item{DAMAGE}{how much damage.}
}
}
\source{
The data was collected from a public government API.
}
\usage{
fire
}
\description{
This dataset contains fire data
}
\examples{
data(fire)
}
\keyword{datasets}
