% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{How many times we want to run}

\item{x}{the data}

\item{fun}{what function we want to estimate}

\item{alpha}{the confidence percentage}

\item{cx}{size of the text}

\item{...}{additional parameters}
}
\value{
a table and a histogram
}
\description{
myboot2
}
\examples{
myboot2(100000,c(1,2,3,4,5,6),fun="mean",alpha=0.05)
}
