% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(sample, conf_interval)
}
\arguments{
\item{sample}{the data to be tested}

\item{conf_interval}{what percentage of confidence}
}
\value{
2 values
}
\description{
myci
}
\examples{
myci(rnorm(30,mean=10,sd=12), conf_interval= .95)
}
