% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Sample size}

\item{iter}{How many times are we running the simulation}

\item{a}{the start of the distribution}

\item{b}{the end of the distribution.}
}
\value{
a grapah and a value w containing the mean of each iteration.
}
\description{
mycltu
}
\examples{
mycltu(1,10000,0,10)
}
