% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function to estimate max like for}

\item{x}{your data}

\item{param}{range of vlaues to eval}

\item{...}{other parameters}
}
\value{
graph and table
}
\description{
mymaxlik
}
\examples{
mymaxlik(lfun = function(x, mu) dnorm(x, mean = mu, sd = 1, log = TRUE),
x = c(-1.2, 0.5, 1.1, 0.8, -0.7, 0.3), param = seq(-2, 2, length.out = 100))
}
