% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{a}{the value up to which the area under the curve is shaded, used to calculate the cumulative probability P(X ≤ a).}
}
\value{
graph and probability
}
\description{
myncurve
}
\examples{
myncurve(mu = 0, sigma = 1, a = 1.5)
myncurve(mu = 5, sigma = 2, a = 6)
}
