% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypiecewise.R
\name{mypiecewise}
\alias{mypiecewise}
\title{Piece wise equation}
\usage{
mypiecewise(x, xk, coef)
}
\arguments{
\item{x}{a vector}

\item{xk}{a int for assigning the break}

\item{coef}{a vector of coeficients of a quadratic linear model.}
}
\value{
a vector
}
\description{
Piece wise equation
}
\examples{
mypiecewise(1,10,100)
}
