% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{mypvalue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{The tcalc value}

\item{xmax}{the max to test}

\item{n}{the size of the sample}

\item{alpha}{the confidence interfval value}
}
\value{
list with p values
}
\description{
mypvalue
}
\examples{
mypvalue(1,n=30,alpha=0.05)
}
