% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{how many items we choose}

\item{iter}{how many samples you want}

\item{time}{how much time we wait}
}
\value{
iter amount of graphs
}
\description{
mysample
}
\examples{
mysample(10)
}
