% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{probability of overbooked}

\item{p}{probability of show}
}
\value{
plots and a table
}
\description{
ntickets
}
\examples{
ntickets(200,.02,.95)
ntickets(400,.02,.95)
}
