% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function to estimate max like for}

\item{x}{your data}

\item{param}{range of vlaues to eval}

\item{...}{other parameters}
}
\value{
graph and table
}
\description{
mymaxlik
}
